extract DisclosureArrow into its own file
