+++ views/components/list/list-row.js
@@ -2,9 +2,8 @@
 import React from 'react'
 import {Platform, StyleSheet, View} from 'react-native'
 import {Touchable} from '../touchable'
+import {DisclosureArrow} from './disclosure-arrow'
 import noop from 'lodash/noop'
-import * as c from './colors'
-import Icon from 'react-native-vector-icons/Ionicons'
 
 const styles = StyleSheet.create({
   container: {
@@ -22,28 +21,6 @@ const styles = StyleSheet.create({
     }),
   },
 })
-
-const arrowStyles = StyleSheet.create({
-  wrapper: {
-    marginLeft: 10,
-  },
-  icon: {
-    color: c.iosDisabledText,
-    fontSize: 20,
-  },
-})
-
-const DisclosureArrow = ({style}: {style?: any}) => {
-  if (Platform.OS === 'android') {
-    return null
-  }
-
-  return (
-    <View style={[arrowStyles.wrapper, style]}>
-      <Icon style={arrowStyles.icon} name='ios-arrow-forward' />
-    </View>
-  )
-}
 
 type PropsType = {
   children?: any,
